style(settings): silence -Wsign-compare in Setting_window2
The line-navigation bound `select < line_total - 1` compared a u32 against a
promoted int. Cast the (always positive) bound to u32. Codegen-neutral: the
implicit usual-arithmetic conversion already made the comparison unsigned, so
the .gba hash is unchanged (verified).

Touched as part of the save-backup work per docs/CODE_QUALITY.md
("fix warnings as we touch the code").

+++ source/setwindow2.c
@@ -280,7 +280,7 @@ u32 Setting_window2(void)
 			}
 			else if (keys & KEY_DOWN)
 			{
-				if (select < line_total - 1)
+				if (select < (u32)(line_total - 1))
 				{
 					select++;
 					re_show = 1;
